Create london-underground.xml
An XML file with a GEXF representation of the London Underground as a Graph. Watch out since this won't 100% correlate with TfL data about stations - e.g., Edgware Road is one station here, but TfL treats it as 3 in its KML representations of the underground. The names are a bit messy, too. I'm working to clean this up a bit.